+++ tourcount/src/main/java/com/wmstein/tourcount/SettingsLActivity.java
@@ -1,12 +1,9 @@
 package com.wmstein.tourcount;
 
-import android.Manifest;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.view.MenuItem;
 
@@ -19,15 +16,13 @@ import androidx.preference.PreferenceManager;
  * Set the Settings parameters for TourCount
  * Based on SettingsActivity created by milo on 05/05/2014.
  * Adapted for TourCount by wmstein on 2022-05-21,
- * last edited on 2022-05-21
+ * last edited on 2023-05-08
  */
 public class SettingsLActivity extends AppCompatActivity
 {
     SharedPreferences prefs;
     private SharedPreferences.Editor editor;
     
-    final private static int REQUEST_CODE_ASK_PERMISSIONS = 123;
-
     @Override
     @SuppressLint({"CommitPrefEdits", "SourceLockedOrientationActivity"})
     public void onCreate(Bundle savedInstanceState)
@@ -45,16 +40,6 @@ public class SettingsLActivity extends AppCompatActivity
 
         prefs = PreferenceManager.getDefaultSharedPreferences(this);
         editor = prefs.edit(); // will be committed on pause
-
-        // permission to read db
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M)
-        {
-            int hasReadStoragePermission = checkSelfPermission(Manifest.permission.READ_EXTERNAL_STORAGE);
-            if (hasReadStoragePermission != PackageManager.PERMISSION_GRANTED)
-            {
-                requestPermissions(new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, REQUEST_CODE_ASK_PERMISSIONS);
-            }
-        }
     }
 
     @Override